Separate nanogui widgets' management from GLFW management
 Made a ScreenCore class (out of Screen) responsible of managing the
widgets - input and rendering - separated from Screen (which inherits
from ScreenCore) that now is responsible of managing GLFW events and
mapping to widgets’ input/rendering. Who desires to manage GLFW events
with custom callbacks (and not with Screen) can derive from ScreenCore
and call directly its callbacks.
