+++ include/nanogui/screen.h
@@ -13,7 +13,7 @@
 
 #pragma once
 
-#include <nanogui/widget.h>
+#include <nanogui/screen_core.h>
 
 NAMESPACE_BEGIN(nanogui)
 
@@ -21,9 +21,7 @@ NAMESPACE_BEGIN(nanogui)
  * \brief Represents a display surface (i.e. a full-screen or windowed GLFW window)
  * and forms the root element of a hierarchy of nanogui widgets
  */
-class NANOGUI_EXPORT Screen : public Widget {
-    friend class Widget;
-    friend class Window;
+class NANOGUI_EXPORT Screen : public ScreenCore {
 public:
     /// Create a new screen
     Screen(const Vector2i &size, const std::string &caption,
@@ -59,12 +57,6 @@ public:
     /// Handle a file drop event
     virtual bool dropEvent(const std::vector<std::string> & /* filenames */) { return false; /* To be overridden */ }
 
-    /// Default keyboard event handler
-    virtual bool keyboardEvent(int key, int scancode, int action, int modifiers);
-
-    /// Text input event handler: codepoint is native endian UTF-32 format
-    virtual bool keyboardCharacterEvent(unsigned int codepoint);
-
     /// Window resize event handler
     virtual bool resizeEvent(const Vector2i &) { return false; }
 
@@ -74,16 +66,9 @@ public:
     /// Return a pointer to the underlying GLFW window data structure
     GLFWwindow *glfwWindow() { return mGLFWWindow; }
 
-    /// Return a pointer to the underlying nanoVG draw context
-    NVGcontext *nvgContext() { return mNVGContext; }
-
     void setShutdownGLFWOnDestruct(bool v) { mShutdownGLFWOnDestruct = v; }
     bool shutdownGLFWOnDestruct() { return mShutdownGLFWOnDestruct; }
 
-    /// Compute the layout of all widgets
-    void performLayout() {
-        Widget::performLayout(mNVGContext);
-    }
 public:
     /********* API for applications which manage GLFW themselves *********/
 
@@ -105,39 +90,16 @@ public:
     void initialize(GLFWwindow *window, bool shutdownGLFWOnDestruct);
 
     /* Event handlers */
-    bool cursorPosCallbackEvent(double x, double y);
-    bool mouseButtonCallbackEvent(int button, int action, int modifiers);
-    bool keyCallbackEvent(int key, int scancode, int action, int mods);
-    bool charCallbackEvent(unsigned int codepoint);
     bool dropCallbackEvent(int count, const char **filenames);
-    bool scrollCallbackEvent(double x, double y);
     bool resizeCallbackEvent(int width, int height);
 
-    /* Internal helper functions */
-    void updateFocus(Widget *widget);
-    void disposeWindow(Window *window);
-    void centerWindow(Window *window);
-    void moveWindowToFront(Window *window);
-    void drawWidgets();
-
-    void performLayout(NVGcontext *ctx) {
-        Widget::performLayout(ctx);
-    }
-
 protected:
+    /// Reimplementing this for calling glfwSetCursor()
+    void setCursorGLFW(int c);
+    
     GLFWwindow *mGLFWWindow;
-    NVGcontext *mNVGContext;
     GLFWcursor *mCursors[(int) Cursor::CursorCount];
-    Cursor mCursor;
-    std::vector<Widget *> mFocusPath;
     Vector2i mFBSize;
-    float mPixelRatio;
-    int mMouseState, mModifiers;
-    Vector2i mMousePos;
-    bool mDragActive;
-    Widget *mDragWidget = nullptr;
-    double mLastInteraction;
-    bool mProcessEvents;
     Vector3f mBackground;
     std::string mCaption;
     bool mShutdownGLFWOnDestruct;
